+++ src/m_cheat.c
@@ -188,6 +188,8 @@ static void cheat_saitama(void);     // MDK Fist
 static void cheat_boomcan(void);     // Explosive hitscan
 
 static void cheat_fauxdemo(void); // Emulates demo/net-play state, for debugging
+static void cheat_netgame(void);
+static void cheat_dmatch(void);
 static void cheat_myrng(void);
 static void cheat_myrngx(char *buf);
 static void cheat_dimlight(void);
@@ -467,13 +469,15 @@ cheat_sequence_t cheats_table[] = {
   { "saitama",    NULL, not_net | not_demo, {.v = cheat_saitama} }, // --------------------------- MDK Fist
   { "boomcan",    NULL, not_net | not_demo, {.v = cheat_boomcan} }, // --------------------------- Explosive hitscan
 
-  { "fauxdemo",   NULL, devmode_only, {.v = cheat_fauxdemo} }, // Emulates demo/net-play state, for debugging
-  { "myrng",      NULL, devmode_only, {.v = cheat_myrng} },
-  { "myrng",      NULL, devmode_only, {.s = cheat_myrngx}, -3 },
-  { "dimlight",   NULL, devmode_only, {.v = cheat_dimlight} },
-  { "fovsky",     NULL, devmode_only, {.v = cheat_fovsky} },
-  { "hiresgfx",   NULL, devmode_only, {.v = cheat_hiresgfx}, .repeatable = true },
-  { "castcall",   NULL, devmode_only, {.v = cheat_castcall} },
+  { "fauxdemo",   NULL, devmode_only|not_demo|not_net, {.v = cheat_fauxdemo} },
+  { "netgame",    NULL, devmode_only|not_demo,         {.v = cheat_netgame}, .repeatable = true },
+  { "dmatch",     NULL, devmode_only|not_demo,         {.v = cheat_dmatch}, .repeatable = true },
+  { "myrng",      NULL, devmode_only|not_demo|not_net, {.v = cheat_myrng} },
+  { "myrng",      NULL, devmode_only|not_demo|not_net, {.s = cheat_myrngx}, -3 },
+  { "dimlight",   NULL, devmode_only,                  {.v = cheat_dimlight}, .repeatable = true },
+  { "fovsky",     NULL, devmode_only,                  {.v = cheat_fovsky}, .repeatable = true },
+  { "hiresgfx",   NULL, devmode_only,                  {.v = cheat_hiresgfx}, .repeatable = true },
+  { "castcall",   NULL, devmode_only|not_demo|not_net, {.v = cheat_castcall} },
 
   { "cheese",     NULL, not_net | not_demo, {.v = cheat_cheese} },
   { "flakes",     NULL, not_net | not_demo, {.v = cheat_flakes} },
@@ -966,6 +970,18 @@ static void cheat_fauxdemo(void)
 
   S_StartSound(plyr->mo, sfx_tink);
   displaymsg("Fauxdemo %s", fauxdemo ? "ON" : "OFF");
+}
+
+static void cheat_netgame(void)
+{
+  netgame = !netgame;
+  displaymsg("Netgame State %s", netgame ? "ON" : "OFF");
+}
+
+static void cheat_dmatch(void)
+{
+  deathmatch = !deathmatch;
+  displaymsg("Deathmatch State %s", deathmatch ? "ON" : "OFF");
 }
 
 static void cheat_myrng(void)
@@ -1979,7 +1995,7 @@ static boolean CheatAllowed(cheat_when_t when)
            && !(when & not_demo && (demorecording || demoplayback))
            && !(when & not_menu && menuactive)
            && !(when & beta_only && !beta_emulation)
-           && !(when & devmode_only_bit && !nugget_devmode); // [Nugget]
+           && !(when & devmode_only && !nugget_devmode); // [Nugget]
 }
 
 // The cheat detection function was replaced with a version from Chocolate Doom