+++ src/st_stuff.c
@@ -3499,10 +3499,13 @@ static void DrawNughudGraphics(void)
   {
     if (weaponinfo[plyr->readyweapon].ammo != am_noammo)
     {
+      const int maxammo_divisor = nughud.ammobar_resize
+                                ? 1 : (1 + plyr->backpack);
+
       DrawNughudBar(
         &nughud.ammobar, nhambar,
         plyr->ammo[weaponinfo[plyr->readyweapon].ammo],
-        plyr->maxammo[weaponinfo[plyr->readyweapon].ammo] / (1 + plyr->backpack)
+        plyr->maxammo[weaponinfo[plyr->readyweapon].ammo] / maxammo_divisor
       );
     }
 
@@ -3510,7 +3513,9 @@ static void DrawNughudGraphics(void)
                armor = sbar_armor;
 
     DrawNughudBar(&nughud.healthbar, nhhlbar, health, deh_max_health);
-    DrawNughudBar(&nughud.armorbar,  nharbar,  armor, deh_max_armor/2);
+
+    if (!nughud.armor_hide || plyr->armorpoints >= 1)
+    { DrawNughudBar(&nughud.armorbar,  nharbar,  armor, deh_max_armor/2); }
   }
 
   for (int i = NUMNUGHUDPATCHES/2;  i < NUMNUGHUDPATCHES;  i++)
@@ -3641,7 +3646,7 @@ static void DrawNughudGraphics(void)
     if (patch) { DrawNughudPatch(&nughud.healthicon, patch, no_offsets); }
   }
 
-  if (nughud.armoricon.x > -1)
+  if (nughud.armoricon.x > -1 && (!nughud.armor_hide || plyr->armorpoints >= 1))
   {
     patch_t *patch = NULL;
     boolean no_offsets = false;
@@ -4312,7 +4317,19 @@ end_amnum:
 
   if (nughud.armor.x > -1)
   {
-    array_push(sb.children, CreateNughudNumber(nughud.armor, sbn_armor, &tnum, 3, true));
+    sbarelem_t elem = CreateNughudNumber(nughud.armor, sbn_armor, &tnum, 3, true);
+
+    if (nughud.armor_hide)
+    {
+      sbarcondition_t condition = {0};
+
+      condition.condition = sbc_armorgreaterequal;
+      condition.param = 1;
+
+      array_push(elem.conditions, condition);
+    }
+
+    array_push(sb.children, elem);
   }
 
   // Ammos -------------------------------------------------------------------